Intermediate commit: revising MIDISequencer and AdvancedSequencer classes.
The MIDISequencer is now a MIDITick class, so I must make it capable to play and
work in a thread safe way. So I  think it should become a very basic sequencer type,
porting all the message processing to the AdvancedSequencer class. This is still unfinished
now, but I want to save the situation.
Main changes:
- Revised documentation
- Revised MIDIManager and MIDIThru eliminating old references in MIDIManager
- Trying to make standard threading behavior of all MIDITick components, using the
  mutex proc_lock for functions changing the parameters
- Revision of MIDISequencer and AdvancedSequencer (in progress)
- Added Metronome and MIDIRecorder class in their own files with
  test_metronome.cpp and test_recorder.cpp files (unfinished)