Second attempt to correct bug in autostop
The last commit did not resolved the bug: when the Sequencer::Stop() is called by autostop it runs in a separate thread and can cause data races. This is the second attempt to fix it. I added also the new example file test_advancedsequencer_noinput.cpp which reproduces (now no more) the bug