Revised many types
I assumed all chars to be signed but this led to errors in some system. So I made a general revision of the types changing many of them. This involves also some function signature, however in a compatible way (many types changed
 from char to int).
Many files are affected.
So I had also to recompile doxygen doc