Attempt to limit the threads
In this version the MIDITimer is started at the beginning by the MIDIManager and never stopped, moreover the sequencer autostop is modified and no more starts a third thread.
So ther are always two thread active: the principal and the timer thread, no trouble  with start/join/detach.
Files affected: manager.cpp, sequencer.cpp (for the TickProc), timer.cpp (only for logs)